requirements.txt
astroid==2.0.4
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.1.1
PyYAML==3.13
ruamel.yaml==0.15.61
six==1.11.0
typed-ast==1.1.0
vext==0.7.0
vext.gi==0.7.0
wrapt==1.10.11
